requirements.txt
openai
numpy
python-telegram-bot
python-dotenv
pymongo